refactor for a single types.py with conditional definitions
