+++ test-integration/test_integration/test_predict.py
@@ -86,11 +86,7 @@ def test_predict_writes_multiple_files_to_files(tmpdir_factory):
     out_dir = pathlib.Path(tmpdir_factory.mktemp("project"))
     shutil.copytree(project_dir, out_dir, dirs_exist_ok=True)
     result = subprocess.run(
-        [
-            "cog",
-            "predict",
-            "--debug"
-        ],
+        ["cog", "predict", "--debug"],
         cwd=out_dir,
         check=True,
         capture_output=True,
